chore: move separate parts of timeline back to one file
